# DBZ-5071 Improve incremental snapshot query performance for PostgreSQL