This branch builds on the suggested code changes made by @jpechane and the query changes suggested by myself in [DBZ-5071](https://issues.redhat.com/browse/DBZ-5071) to improve the incremental snapshot query performance for PostgreSQL by using row value constructors.

@jpechane and I also noted and discussed special considerations around NULL handling, and one thing I realized is that the connector doesn't do well with rows that have NULL value in their keys today.  The connector might crash, or it might simply omit them entirely.

This branch therefore has three main commits:

- Refactor some functions into the suggested ChunkQueryBuilder interface.
- Fix all the erroneous NULL handling by making the default chunk query builder NULL-aware.
- Create RowValueConstructorChunkQueryBuilder and use it for PostgreSQL.

Each is described in more detail in the respective commit message.  The changes are dependent upon each other and thus in one branch.  If the Debezium maintainers prefer, I could split this into three sequential pull requests instead (let me know).  The second commit in particular is arguably scope-creeping DBZ-5071, itself an extension of the scope-creeping discussion in DBZ-5071 about NULL handling.

It was suggested in the JIRA issue to also apply this for MySQL.  However, that is counterproductive because they don't [support taking advantage of them in the query planner](https://bugs.mysql.com/bug.php?id=111952).  In fact, with some [simple experimentation](https://onecompiler.com/mysql/3zzdsudzx) I found that MySQL is already using optimal queries using the existing queries generated by the default chunk query builder.  Therefore, I suggest we not mess with MySQL further unless there are complaints about performance that can be traced to poor query plans.  The original complaint in DBZ-5071 was about PostgreSQL.  The good news is that with this refactor, it's easy to update MySQL connector in the future to use row value constructors if/when MySQL itself can utilize them properly.

Test results on a table with a ~few million rows (subset of real production data), run on a local PostgreSQL 13 instance, local Kafka broker, local Connect instance:

- Old incremental scanning code: durations:
    - 1st run: 0.51s user 0.26s system 0% cpu 1:50.70 total (cold Connect restart)
    - 2nd run: 0.48s user 0.20s system 0% cpu 1:45.67 total
    - 3rd run: 0.50s user 0.20s system 0% cpu 1:45.64 total
- New incremental scanning code: durations:
    - 1st run: 0.60s user 0.28s system 1% cpu 1:13.49 total (cold Connect restart)
    - 2nd run: 0.48s user 0.18s system 1% cpu 1:02.84 total
    - 3rd run: 0.49s user 0.16s system 1% cpu 1:02.47 total
- Ad-hoc blocking snapshot: durations:
    - 1st run: 0.44s user 0.17s system 1% cpu 58.627 total
    - 2nd run: 0.43s user 0.17s system 1% cpu 60.000 total
    - 3rd run: 0.42s user 0.16s system 0% cpu 58.959 total

I noticed that the incremental and ad-hoc blocking snapshot times were suspiciously close together.  `docker stats` reports that PG only uses minimal CPU in the last two cases, and Connect is clearly CPU-bound.  Spot checking with VisualVM and I found multiple hotspots and potential areas for improvement, but that's another story.  The point is: the time savings with this pull request is probably even greater than the numbers above, because now it's Connect's CPU usage limiting performance instead of PostgreSQL query performance.