@james-johnston-thumbtack 2024-01-15
The MongoDB 5.0 failure (`MultiThreadIncrementalSnapshotIT#pauseDuringSnapshot`) seems like an unrelated flaky test to me.  I tried running it myself and the test passes: `debezium/debezium-connector-mongodb$ ../mvnw clean verify -DskipTests -DskipITs=false -Dit.test=io.debezium.connector.mongodb.MultiThreadIncrementalSnapshotIT#pauseDuringSnapshot -Dversion.mongo.server=5.0`  (verified it was MongoDb 5.0 by checking Docker logs while test ran)

The MySQL percona failure (`SnapshotParallelSourceIT#shouldCreateSnapshotOfSingleDatabaseWithoutGlobalLock`) also I suspect to be unrelated flaky.  I similarly couldn't reproduce it locally /w percona docker container running.

SQL Server failures look to be a similar situation:
```
2024-01-15T14:40:06.6588694Z [ERROR]   IncrementalSnapshotWithRecompileIT.stopCurrentIncrementalSnapshotWithAllCollectionsAndTakeNewNewIncrementalSnapshotAfterRestart:176->AbstractIncrementalSnapshotTest.stopCurrentIncrementalSnapshotWithAllCollectionsAndTakeNewNewIncrementalSnapshotAfterRestart:596->AbstractSnapshotTest.startConnector:288->AbstractSnapshotTest.startConnector:279->AbstractConnectorTest.assertNoRecordsToConsume:983 expected:<[tru]e> but was:<[fals]e>
2024-01-15T14:40:06.6593135Z [ERROR]   IncrementalSnapshotWithRecompileIT>AbstractIncrementalSnapshotTest.stopCurrentIncrementalSnapshotWithoutCollectionsAndTakeNewNewIncrementalSnapshotAfterRestart:541->AbstractSnapshotTest.startConnector:288->AbstractSnapshotTest.startConnector:279->AbstractConnectorTest.assertNoRecordsToConsume:983 expected:<[tru]e> but was:<[fals]e>
2024-01-15T14:40:06.6596091Z [ERROR]   SqlServerReselectColumnsProcessorIT.afterEach:41->AbstractReselectProcessorTest.afterEach:67->AbstractConnectorTest.assertNoRecordsToConsume:983 expected:<[tru]e> but was:<[fals]e>
```
The first one is explicitly tagged Flaky, but is simply a passthrough that calls into the super class, which is not designated flaky (yet also failed).  The last one is not designated flaky, but also appears unrelated to this branch.  Given that the tests are either marked flaky or seem unrelated, I'm inclined to suspect none of the failures have anything to do with this PR.

For spanner, this test failed seemed unrelated: `io.debezium.connector.spanner.kafka.internal.BufferedPublisherTest.testBufferedPublisher_2`