@james-johnston-thumbtack 2024-01-15
**The Db2 failure, however, is a compilation failure and raises some question how best to proceed**.  Would appreciate some guidance here, especially since that connector is in a separate repo.  The issue arises from the second commit in this branch that improves NULL handling.  The following changes in that commit seem notable that could cause test failures in derived connectors:
- The `AbstractIncrementalSnapshotTest` class had two abstract functions that provide table name, and two tests added to it related to testing incremental snapshots with NULL keys.  This was motivated due to the copy/pasted code between PostgreSQL and MySQL, and I didn't want to further copy/paste into Oracle and SQL Server.  I also didn't want to leave those connectors with a gap in test coverage (which turned out to be needed).  Derived classes have to create table schema for the test (if they weren't already), set up message key column connector setting, and override those abstract functions.  But now the Db2 connector fails because it derives its own incremental snapshot IT from this abstract class.
- The `JdbcConnector` has a new `nullsSortLast` function which is used by the default chunk builder to build the NULL-aware chunking queries.  Although the function is not abstract, it does return a tri-state `Optional<Boolean>` and the base class returns empty, meaning "I don't know" (there's no reasonable default to assume here).  If this is empty value is found by the query chunk builder, _and_ a key column is nullable, then the query chunk builder now fails fast with an error.  This is arguably an improvement over older Debezium versions which, while they didn't fail, would _silently omit records from the snapshot_ if there were NULL values.  In this PR, I override that function in every connector (PG, MySQL, Oracle, SQL Server) with the correct value based on the vendor's documentation.  But connectors outside this repo would also need updating.

One open question in my mind is what level of compatibility needs to be guaranteed.  For example, how stable are the test cases in the `AbstractIncrementalSnapshotTest` really intended to be / are they set in stone or can we make changes like I'm proposing to do here?

I am happy to try making the improvements to the Db2 (and any other) connectors as well, but would appreciate some feedback on the best way to proceed before I spend more time on it.  Here's one idea:

1.  Split this PR into three or more.  I feel more like this PR might be biting off more than it can chew.
2.  The first PR merges the QueryChunkBuilder refactor, because it's relatively straightforward, yet also touches a number of different places that might be more likely to result in a merge conflict as time passes.
3.  Other connector repos are prepared by ensuring the integration tests create tables with NOT NULL columns.  If they had duplicated the code from PG/MySQL that is being abstracted into the base test class, then those tests are deleted.  In practice we figure out what changes to make here by observing the GitHub Actions workflow failures with the next step, below.
4.  The second PR is the same as the "Correctly handle NULL values" commit from this PR, except that the new functions and tests in `AbstractIncrementalSnapshotTest` _are not abstract_.  Instead, a new (temporary) function is also added that allows derived classes to opt-in to the new nullability tests.  The default behavior is to opt-out.  The tests will be no-ops in such an opt-out case.
5.  The other connector repos are then modified to override the new `nullsSortLast` function in `JdbcConnector`.  The tests are also updated to opt-in to the new nullable key tests in the base class by overriding the appropriate functions.
6.  If we want, at this point, we can clean up the `AbstractIncrementalSnapshotTest` a bit by making the functions abstract as currently being proposed in this PR, rather than returning placeholder/invalid defaults.  The "opt-in/opt-out" function is changed to return "opt-in" by default.  This is merged, and then the connectors are updated to remove the override of the opt-in/out function.  Finally, the opt-in/out function is deleted entirely.
7.  After all this, merge the 3rd PR setting up the RowValueConstructorQueryChunkBuilder; this should also be straightforward since the PG connector is in the same repo as core packages.

A variation of the idea could also use the opportunity to make `JdbcConnection` and the new `nullsSortLast` both abstract, if we wanted to force derived classes to provide this information at compile-time, rather than waiting until runtime to fail.  I guess it comes down to: what compatibility guarantee needs to be made here...?