@jpechane 2024-01-25
@james-johnston-thumbtack Thank you very much for the PR. Generally it LGTM but I need to take a look in more detail.

WRT how to continue, I don't think it is necessary to split the PR. Yes, it is a bit large but it is quite easy to understand. In relation to Db2 - I think it is just enough to create a PR in Db2 repo that will use the branch with the same name. In that case Db2 PR will be built and tested against this PR. Just please be aware that DB2 CI is a bit unstable in relation to incremental snapshots so please make sure you run it locally and report the results from local run too so we'll avoid back and forth pinging.