@james-johnston-thumbtack 2024-01-27
Hi @jpechane - that sounds much simpler to proceed!  I prepared a PR as recommended for the Db2 connector: https://github.com/debezium/debezium-connector-db2/pull/131.  The incremental tests pass when run locally, as reported in my comment there.